+++ RELEASES.md
@@ -5,9 +5,17 @@
 - replace brownian movement with groc-seeking-groc behavior
 - parameterize 
 
-### world (v0.3.0) -- in progress
+### world (v0.3.0) -- complete
 
 - separate Groc class and World class
 - import Groc (into World)
-- behavior/motion/motivation moved to Groc class
 - environment/movement/location moved to World class
+
+### auto (v0.4.0) -- complete
+
+- behavior/motion/motivation moved to Groc class
+- fix Groc/World interaction regarding movement
+- add birthTick (groc world time)
+- remove birthdatetime (gregorian)
+- add gender
+- augment groc.dat file format